Implement a simple 'keystore' plugin
This implements a simple 'keystore' plugin, which allows to persist key-value
pairs. The store is attached to the bot and is used in the following way:

  local bot = ... -- create a bot
  bot:add_plugin("keystore", { backend = "filesystem", path = "./keystore" })
  bot.keystore:set("some-value", "Hello, keystore")
  print(bot.keystore:get("some-value"))

Currently the 'filesystem' backend does not perform any escaping at all of
the file names: it just uses the keys as-is.

This is an initial attempt at tackling #5, but the fact that this initial has
many shortcomings makes it not enough to close the issue.