meeting: Templates for messages, improve data representation
This commits changes the representation of the meeting log items, as to split
them between raw text "log" items (everything that is said in the room while
a meeting is held) and "minutes" items (items which go in the minutes).
Formatting of messages is left to a mini template language, instead of
printf-style formats, which would better allow later on for localization
(if ever implemented) or overriding messages from the configuration.

As a bonus, this also implements the #undo and #nick commands, which
are trivial now with the new data representation, contributing to the
completion of issue #2.