Ensure that plugins are loaded only once
This keeps a "bot.plugin[name]" table, with the value returned by the plugin
initialization function (if any), or a truthy value. Plugins can store the
data they need for their operation in "bot.plugin[name]" (which usually will
be a table), and it the bot framework in general will treat it as private
data.

This also modifies two existing plugins to make use of the new loading
mechanism:

- Plugin "keystore" sets "bot.plugin.keystore" to a table with two methods:
  :get() and :set(), which work as expected.
- Plugin "meeting" tracks active meetings in "bot.plugin.meeting".