Keep configuration in "bot.config" at runtime
Instead of passing the local/global configuration around to each plugin
initializer, keep the settings in a "bot.config" table.

For now this allows to move joining rooms to "plugin/muc.lua", and avoid the
kludge of manually loading the "muc" plugin if rooms are configured. Instead,
the configured rooms are moved by the bot:configure() function into the config
of the "muc" plugin, and the plugin itself hooks in its initializer to the
"started" event in order to join the preconfigured rooms.

Also, plugins can now check each others' cofiguration (if needed), and in the
future the bot may notify plugins about setting changes. Which in turn would
allow for dynamic reconfiguration. But that's getting way ahead of ourselves.