commandevent: Do not require (semi)colon/comma to address bot
This allows to mention commands using the bot nick in a MUC room without
needing to add the trailing comma/semicolon/colon after the bot nickname.
A space is still required, though. With this patch, now all the following
ways of addressing the bot work:

  botnick: command args
  botnick; command args
  botnick, command args
  botnick command args

Note that the string is still matched at the beginning of the message.