commandevent: Implement a subcommand dispatch decorator
This patch implements a "dispatch()" function, useable from other plugins as
"<bot>.plugin.commandevent.dispatch()", which takes a table of functions and
dispatches subcommands to different functions. For example:

  bot:hook("command/hello", bot.plugin.commandevent.dispatch {
    peter = function () return "/say Peter was greeted" end;
    world = function () return "/say The World was greeted" end;
    _     = function () return "Who would you like to greet?" end;
  })

Note that the underscore ("_") is the "wildcard" subcommand, and it will be
used when the no valid subcommand has been provided.

This will help implementing e.g. issue #12.