httpevent: Support exporting functionality over HTTP
The "httpevent" plugin allows luabot to listen over HTTP, and react
to incoming web requests. For example, another plugin could use this
functionality to listen to webhooks and notifications to a MUC room.

In order to enable the HTTP server functionality, the "httpevent"
plugin *must* be configured with (at least) the listening port:

    plugin "httpevent" {
	port = 8888,        -- Mandatory
	host = "localhost"  -- Optional
    }

Other plugins can make use of the new "http/<method>/<uri>" events:

    bot:hook("http/post/echo", function (event)
	-- Print the received HTTP headers.
	for name, value in pairs(event.request.headers) do
	    print("Header: " .. name .. ": " .. value)
	end

	-- Print the HTTP request body.
	print(event.request.body)

	-- Return "200 OK" to the client, plus some text.
	-- To generate a different HTTP response, return
	-- it numeric code instead e.g. "return 401", or
	-- use "event.response" and its methods.

	return "Response body text"
    end)

The above can be accessed using a tool like curl in the following way:

    curl -s --data "Hello, world!" http://localhost:8888/echo

This patch closes #9.