Make strutil.interpolate resolve dotted paths
This makes strutil.interpolate() support "dotted paths" in variable
expansions. This allows template strings to obtain values from nested
tables:

    strutil.interpolate("%{foo.bar}", { foo = { bar = 42 } })

evaluates to "42". If any of the intermediate values resolved to "nil",
or a value cannot be indexed (e.g. is "foo" above wouldn't be a table)
then the expansion is not performed:

    strutil.interpolate("%{foo} %{foo.bar.baz}", { foo = true })

evaluates to "true %{foo.bar.baz}" because the dotted path in the second
expansion cannot be resolved.