Modify package.path to load modules installed by LuaRocks
When luabot is installed using LuaRocks, an additional "luabot" path component
is prepended to modules, but modules are require()'d without it as a prefix.
This modifies package.path to add the prefix in the search path, so the
require() calls do not need to be changed. This is particularly important
to avoid the need to modify all the require() calls in third party modules
imported into the luabot source tree (verse and its dependencies, mainly).